+++ xarray/core/treenode.py
@@ -468,7 +468,7 @@ class TreeNode(Generic[Tree]):
                     current_node = current_node.get(part)
         return current_node
 
-    def _set(self: Tree, key: str, val: Tree) -> None:
+    def _set(self: Tree, key: str, val: Tree, *, copy: bool = True) -> None:
         """
         Set the child node with the specified key to value.
 
@@ -483,6 +483,8 @@ class TreeNode(Generic[Tree]):
         item: Tree | T_DataArray,
         new_nodes_along_path: bool = False,
         allow_overwrite: bool = True,
+        *,
+        copy: bool = True,
     ) -> None:
         """
         Set a new item in the tree, overwriting anything already present at that path.
@@ -500,6 +502,8 @@ class TreeNode(Generic[Tree]):
         allow_overwrite : bool
             Whether or not to overwrite any existing node at the location given
             by path.
+        copy : bool, optional
+            Whether to make a shallow copy of the values in the node. Default is True.
 
         Raises
         ------
@@ -539,7 +543,7 @@ class TreeNode(Generic[Tree]):
                     elif new_nodes_along_path:
                         # Want child classes (i.e. DataTree) to populate tree with their own types
                         new_node = type(self)()
-                        current_node._set(part, new_node)
+                        current_node._set(part, new_node, copy=copy)
                         current_node = current_node.children[part]
                     else:
                         raise KeyError(f"Could not reach node at path {path}")
@@ -547,11 +551,11 @@ class TreeNode(Generic[Tree]):
         if name in current_node.children:
             # Deal with anything already existing at this location
             if allow_overwrite:
-                current_node._set(name, item)
+                current_node._set(name, item, copy=copy)
             else:
                 raise KeyError(f"Already a node object at path {path}")
         else:
-            current_node._set(name, item)
+            current_node._set(name, item, copy=copy)
 
     def __delitem__(self: Tree, key: str):
         """Remove a child node from this tree object."""