+++ doc/whats-new.rst
@@ -22,6 +22,9 @@ v2024.06.1 (unreleased)
 
 New Features
 ~~~~~~~~~~~~
+- Add optional parameter `copy` to :py:func:`xarray.core.datatree.DataTree.from_dict`,
+  which by default shallow copies the values in the dict. (:pull:`9193`)
+  By `Jimmy Westling <https://github.com/illviljan>`_.
 - Allow chunking for arrays with duplicated dimension names (:issue:`8759`, :pull:`9099`).
   By `Martin Raspaud <https://github.com/mraspaud>`_.
 - Extract the source url from fsspec objects (:issue:`9142`, :pull:`8923`).